Revert "Simple tracker draft"
