--- examples/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 3.5)
-
-add_subdirectory(calibration_load)
-add_subdirectory(calibration_store)
-add_subdirectory(detector_and_tracker)
-add_subdirectory(distance)
