--- examples/detector_and_tracker/detector_and_tracker_example.cpp
@@ -1,120 +1,0 @@
-/*****************************************************************//**
- * @file   detector_and_tracker_example.cpp
- * @brief  Example that detects Kilobots in video file and then tracks them.
- *         Displays IDs of detected Kilobots.
- *
- * @author Petr Svoboda
- * @date   May 2022
-***********************************************************************/
-
-#include <opencv2/imgproc.hpp>
-#include <opencv2/highgui.hpp>
-
-#include "kilobot.h"
-#include "euclidian_tracker.h"
-#include "yolo_detector.h"
-#include "kilolib_functions.h"
-#include "simpletracker.h"
-
-using namespace kilolib;
-using namespace cv;
-using namespace std;
-
-int main(int argc, char** argv)
-{
-    // Creates YOLO detector and loads DNN model "best.onnx"
-    YoloDetector detector;
-    if (YD_RESULT::YD_OK != detector.LoadNet("../../../utils/best.onnx", false)){
-        return -1;
-    }
-
-    // Open video "bodyguard.mp4"
-    cv::VideoCapture capture("../../../utils/bodyguard.mp4");
-
-    if (!capture.isOpened())
-    {
-        std::cerr << "Error opening video file\n";
-        return -1;
-    }
-
-    // default values 
-    int nms = 0;
-    int score = 5;
-    int conf = 7;
-    int t = 200;
-    int per = 50;
-
-    //create a window called "Controls" with trackbars
-    namedWindow("Controls", WINDOW_AUTOSIZE); 
-    createTrackbar("NMS", "Controls", &nms, 10);
-    createTrackbar("Score", "Controls", &score, 10);
-    createTrackbar("Confidence", "Controls", &conf, 10);
-    createTrackbar("Threshold", "Controls", &t, 255);
-    createTrackbar("Percentage", "Controls", &per, 50);
-
-    // vectors, where detections on current and previous frame will be stored
-    std::vector<Kilobot> output;
-    std::vector<Kilobot> prevOutput;
-
-    cv::Mat frame;
-
-    EuclidianTracker tracker;
-
-    SimpleTracker tr;
-
-    while (true) {
-        // read frame
-        capture.read(frame);
-
-        if (frame.empty()) {
-            std::cout << "End of stream\n";
-            break;
-        }
-
-        // detect Kilobots
-        auto start = std::chrono::steady_clock::now();
-        detector.Detect(frame, output, (float) (nms + 1) / 10, (float) (conf + 1) / 10, (float) (score + 1) / 10);
-        auto stop = std::chrono::steady_clock::now();
-        std::chrono::duration<double> elapsed_seconds = stop - start;
-        std::cout << "NN elapsed time: "
-                  << std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(elapsed_seconds).count()
-                  << "ms\n";
-
-
-        // track Kilobots
-        start = std::chrono::steady_clock::now();
-        tracker.Track(frame, output, prevOutput, (double) t + 1, per / 10);
-        stop = std::chrono::steady_clock::now();
-        elapsed_seconds = stop - start;
-        std::cout << "Euclidian tracker elapsed time: "
-                  << std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(elapsed_seconds).count()
-                  << "ms\n";
-
-        start = std::chrono::steady_clock::now();
-        tr.processKilobots(output);
-        stop = std::chrono::steady_clock::now();
-        elapsed_seconds = stop - start;
-        std::cout << "My tracker elapsed time: "
-                  << std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(elapsed_seconds).count()
-                  << "ms\n";
-
-        tr.drawKilobots(frame);
-        // show output
-        imshow("output", frame);
-
-        if (cv::waitKey(1) != -1) {
-            capture.release();
-            std::cout << "finished by user\n";
-            break;
-        }
-
-        // set current detections to prevOutput for calculations on next frame
-        prevOutput = output;
-        // clear current detections
-        output.clear();
-    }
-
-    cv::waitKey(0);
-
-    return 0;
-}
