+++ include/yolo_detector.h
@@ -11,53 +11,48 @@
 
 #include <opencv2/opencv.hpp>
 #include <opencv2/dnn.hpp>
-#include <string>
-#include <vector>
 
 #include "kilobot.h"
 
+using namespace std;
+using namespace cv;
 
-static const int INPUT_WIDTH = 640;
-static const int INPUT_HEIGHT = 640;
+static const float INPUT_WIDTH = 640.0;
+static const float INPUT_HEIGHT = 640.0;
 
 namespace kilolib {
 
-    /**
-     * @brief Result data type for functions from YoloDetector class
-     */
-    enum class YD_RESULT {
-        YD_OK,          /**< correct function execution */
-        YD_ARGS_ERROR,  /**< unexpected or incorrect input parameters */
-        YD_ERROR        /**< unspecified error */
-    };
-
     class YoloDetector {
     public:
-
-        explicit YoloDetector(const std::string &fileName = "", bool is_cuda = false);
-
-        YD_RESULT LoadNet(const std::string &pathToFile, bool is_cuda);
-
-        YD_RESULT
-        Detect(const cv::Mat &frame, std::vector<Kilobot> &output, float scoreVal, float confVal, float nmsVal);
+        /**
+         * @brief Constructor
+        */
+        YoloDetector();
 
         /**
-         * @brief Check if some network is loaded or not.
-         * @return True if the network is correctly loaded, false otherwise.
-         */
-        bool isLoaded() const { return !_net.empty(); }
+         * @brief Loads model from path and determinates running on CPU or GPU.
+         *
+         * @param pathToFile Relative path to file with pretrained YOLOV5 model.
+         * @param is_cuda If true, function tries to run Net with CUDA backend.
+        */
+        void LoadNet(string pathToFile, bool is_cuda);
 
-    protected:
-        static YD_RESULT
-        _createInputBlob(const cv::Mat &inFrame, cv::Mat &outBlob, float &outXFactor, float &outYFactor);
+        /**
+         * @brief Method used for Kilobot detection in frame.
+         *
+         * @param frame Frame used for detection.
+         * @param output Vector in which detections will be stored.
+         * @param scoreVal Minimum value for kilobot score. Values greater than this are considered to be Kilobots.
+         * @param confVal Minimum value for confidence used by Net. Values greater than this are considered to be objects.
+         * @param nmsVal Value usedfor non-maximum suppression.
+        */
+        void Detect(cv::Mat& frame, std::vector<Kilobot>& output, float scoreVal, float confVal, float nmsVal);
 
-        static YD_RESULT _parseNNResults(const std::vector<std::vector<Mat>> &nnOutput, std::vector<float> &confidences,
-                                         std::vector<cv::Rect> &boxes, int resultRows, float xScale,
-                                         float yScale, float scoreVal, float confVal);
-
-        static cv::Mat _format(const cv::Mat &source);
+    private:
+        cv::Mat _format(const cv::Mat& source);
 
     private:
         cv::dnn::Net _net; /*!< Loaded DNN Net model. */
+        std::vector<cv::Mat> _outputs; /*!< Output of DNN detection. */
     };
 }
