+++ src/euclidian_tracker.cpp
@@ -11,22 +11,25 @@ namespace kilolib {
     void EuclidianTracker::Track(Mat frame, std::vector<Kilobot>& output, std::vector<Kilobot>& prevOutput, double LEDthreshVal, double LEDPercentageVal)
     {
         std::vector<Kilobot> undetected;
+
         // check every single Kilobot in vector of Kilobots found in current frame
-        for (auto &kilobot: output) {
-            //auto detection = output[o];
-            auto box = kilobot.box;
+        for (int o = 0; o < output.size(); ++o)
+        {
+            auto detection = output[o];
+
+            auto box = detection.box;
 
             // try to identify Kilobot in last detections
-            _findObject(frame, kilobot, prevOutput, LEDthreshVal, LEDPercentageVal);
+            _findObject(frame, output[o], prevOutput, LEDthreshVal, LEDPercentageVal);
 
             // if Kilobot wasn't found in previous detections
-            if (kilobot.id == -1) {
+            if (output[o].id == -1) {
                 // check all Kilobots that weren't detected in previous iterations
-                _findObject(frame, kilobot, _prevUndetected, LEDthreshVal, LEDPercentageVal);
+                _findObject(frame, output[o], _prevUndetected, LEDthreshVal, LEDPercentageVal);
 
                 // if even than Kilobot wasn't found, generate new ID and assign it to Kilobot
-                if (kilobot.id == -1) {
-                    kilobot.id = _lastId;
+                if (output[o].id == -1) {
+                    output[o].id = _lastId;
                     _lastId++;
                 }
             }
